// apps/server/src/domain/room/models/room-state.ts
export enum RoomState {
    VACANT = 'VACANT',
    OCCUPIED = 'OCCUPIED',
    CHECKOUT = 'CHECKOUT',
    DIRTY = 'DIRTY',
    CLEANING = 'CLEANING',
    DEEP_CLEANING = 'DEEP_CLEANING',
    INSPECTION = 'INSPECTION',
    AUDIT = 'AUDIT',
    BLOCKED = 'BLOCKED',
    MAINTENANCE = 'MAINTENANCE',
}

export enum RoomStateEvent {
    CHECK_IN = 'CHECK_IN',
    CHECK_OUT = 'CHECK_OUT',
    CANCEL_CHECKOUT = 'CANCEL_CHECKOUT',
    CHECKOUT_TIMEOUT = 'CHECKOUT_TIMEOUT',
    COMPLETE_RENTAL = 'COMPLETE_RENTAL',
    CANCEL_RENTAL = 'CANCEL_RENTAL',
    TRANSFER = 'TRANSFER',
    PERFORM_CLEANING = 'PERFORM_CLEANING',
    COMPLETE_CLEANING = 'COMPLETE_CLEANING',
    CANCEL_CLEANING = 'CANCEL_CLEANING',
    CLEANING_TIMEOUT = 'CLEANING_TIMEOUT',
    PERFORM_DEEP_CLEANING = 'PERFORM_DEEP_CLEANING',
    COMPLETE_DEEP_CLEANING = 'COMPLETE_DEEP_CLEANING',
    CANCEL_DEEP_CLEANING = 'CANCEL_DEEP_CLEANING',
    DEEP_CLEANING_TIMEOUT = 'DEEP_CLEANING_TIMEOUT',
    APPROVE_INSPECTION = 'APPROVE_INSPECTION',
    REJECT_INSPECTION = 'REJECT_INSPECTION',
    INSPECTION_TIMEOUT = 'INSPECTION_TIMEOUT',
    AUDIT = 'AUDIT',
    COMPLETE_AUDIT_VACANT = 'COMPLETE_AUDIT_VACANT',
    COMPLETE_AUDIT_DIRTY = 'COMPLETE_AUDIT_DIRTY',
    COMPLETE_AUDIT_BLOCKED = 'COMPLETE_AUDIT_BLOCKED',
    AUDIT_TIMEOUT = 'AUDIT_TIMEOUT',
    BLOCK = 'BLOCK',
    UNBLOCK = 'UNBLOCK',
    PERFORM_MAINTENANCE = 'PERFORM_MAINTENANCE',
    COMPLETE_MAINTENANCE = 'COMPLETE_MAINTENANCE',
}